Renamed a few packages in the demo bundles. See openhab/static-code-analysis#100
