+++ bundles/automation/org.eclipse.smarthome.automation.sample.moduletype.demo/src/main/java/org/eclipse/smarthome/automation/sample/moduletype/demo/internal/commands/PostEventCommand.java
@@ -14,7 +14,7 @@
  * Plamen Peev - Bosch Software Innovations GmbH - Please refer to git log
  *
  *******************************************************************************/
-package org.eclipse.smarthome.automation.sample.moduletype.commands;
+package org.eclipse.smarthome.automation.sample.moduletype.demo.internal.commands;
 
 import java.util.HashMap;
 import java.util.Map;
